// src/components/pqrs/types.ts
export interface PQRSFormData {
  tipo_solicitud: string;
  proyecto: string;
  descripcion: string;
  num_cuenta: string;
  identificacion_titular: string;
  nombre_titular: string;
  nombre: string;
  correo: string;
  identificacion: string;
  documento: string;
  direccion: string;
  ciudad: string;
  telefono: string;
  genero: string;
}